# apps/credential-manager/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/credentials": {
      "get": {
        "operationId": "CredentialsController_find",
        "summary": "Fetch a list of credentials",
        "description": "This call provides a list of credentials for a given tenant",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credentials fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credentials fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credentials fetched successfully",
                      "data": [{ "id": "71b784a3" }]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Credentials"]
      }
    },
    "/v1/credentials/{credentialRecordId}": {
      "get": {
        "operationId": "CredentialsController_get",
        "summary": "Fetch a credential",
        "description": "This call provides a credential for a given tenant",
        "parameters": [
          {
            "name": "credentialRecordId",
            "required": true,
            "in": "path",
            "description": "The credential record ID to retrieve",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential fetched successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Credentials"]
      }
    },
    "/v1/credentials/{credentialId}/revoke": {
      "post": {
        "operationId": "CredentialsController_revoke",
        "summary": "Revoke a credential",
        "description": "This call revokes a credential for a given tenant",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential revoked successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential revoked successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential revoked successfully",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Credentials"]
      }
    },
    "/v1/credentials/{credentialId}": {
      "delete": {
        "operationId": "CredentialsController_delete",
        "summary": "Delete a credential",
        "description": "This call deletes a credential for a given tenant",
        "parameters": [
          {
            "name": "credentialRecordId",
            "required": true,
            "in": "path",
            "description": "The credential record ID to delete",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential deleted successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential deleted successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential deleted successfully",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Credentials"]
      }
    },
...
  },
  "info": {
    "title": "Gaia-X OCM Credential Manager",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}